Refactor and reorganize code structure; implement GUI navigation and data parsing enhancements - working state
