+++ .gitignore
@@ -2,4 +2,5 @@
 data/raw/*
 /IDEAS.md
 data/*.json
-*/__pycache__/
+__pycache__/
+*.pyc
